move CodeEditor.vue and Markdown.vue to components/editor
